changed auth logic on backend. removed comments. corrected bug in reviewgrid.jsx

+++ server/controllers/ArtistReviewById.py
@@ -12,10 +12,13 @@ class ArtistReviewById(Resource):
   @jwt_required()
   def get(self, artist_review_id):
     user_id = int(get_jwt_identity())
-    review = ArtistReview.query.filter_by(id=artist_review_id, user_id=user_id).first() # find review by id
+    review = ArtistReview.query.filter_by(id=artist_review_id).first() # find review by id
 
     if not review:
       return {'errors': ['Review not found']}, 404
+    
+    if review.user_id != user_id:
+      return {'errors': ['403 Not authorized']}, 403
 
     return ArtistReviewSchema().dump(review), 200
 
@@ -23,10 +26,13 @@ class ArtistReviewById(Resource):
   @jwt_required()
   def patch(self, artist_review_id):
     user_id = int(get_jwt_identity())
-    review = ArtistReview.query.filter(ArtistReview.id == artist_review_id, ArtistReview.user_id == user_id).first()
+    review = ArtistReview.query.filter(ArtistReview.id == artist_review_id).first()
 
     if not review:
       return {'errors': ['Review not found']}, 404
+
+    if review.user_id != user_id:
+      return {'errors': ['403 Not authorized']}, 403
   
     request_json = request.get_json()
 
@@ -49,10 +55,13 @@ class ArtistReviewById(Resource):
   @jwt_required()
   def delete(self, artist_review_id):
     user_id = int(get_jwt_identity())
-    review = ArtistReview.query.filter(ArtistReview.id == artist_review_id, ArtistReview.user_id == user_id).first()
+    review = ArtistReview.query.filter(ArtistReview.id == artist_review_id).first()
 
     if not review:
       return {'errors': ['Review not found']}, 404
+
+    if review.user_id != user_id:
+      return {'errors': ['403 Not authorized']}, 403
 
     db.session.delete(review)
     db.session.commit()
